src/requirements.txt
cloudshell-shell-core
cloudshell-automation-api